fixed the bug when there's no given file to process

+++ cgi-bin/upload.py
@@ -26,11 +26,12 @@ except ImportError:
 
 UPLOAD_DIR = "tmp"
 IMAGE ='none'
+IMAGE_AVAILABLE = False
 
 HTML_TEMPLATE = """<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
 <html><head><title>File Upload</title>
 <meta http-equiv="Content-Type" content="text/html; charset=iso-8859-1">
-</head><body><h1>File Upload</h1>
+</head><body><h1>Upload face</h1>
 <form action="%(SCRIPT_NAME)s" method="POST" enctype="multipart/form-data">
 File name: <input name="file_1" type="file"><br>
 <input name="submit" type="submit">
@@ -73,15 +74,22 @@ def save_uploaded_file (form_field, upload_dir):
 
 def getDimensions( image ):
     img=cv2.imread(image,0)
-    # print img
-    height, width = img.shape
-    dimensions ={'height': height, 'width': width}
-    dumped=json.dumps(dimensions)
-    print
-    print dumped
+    if((img is not None) and img.any()):
+        global IMAGE_AVAILABLE
+        IMAGE_AVAILABLE=True
+        # print img
+        height, width = img.shape
+        dimensions ={'image':image, 'height': height, 'width': width}
+        dumped=json.dumps(dimensions)
+        print
+        print dumped
+
+    else:
+        print "no image uploaded"
 
 def delete (image):
-     os.remove(image)
+    if(IMAGE_AVAILABLE):
+        os.remove(image)
 
 save_uploaded_file ("file_1", UPLOAD_DIR)
 print_html_form ()
